Moving other to domain package.
