+++ app/src/main/java/com/gb/stopwatch/domain/StopwatchState.kt
@@ -1,4 +1,4 @@
-package com.gb.stopwatch
+package com.gb.stopwatch.domain
 
 sealed class StopwatchState {
