+++ app/src/main/java/com/gb/stopwatch/domain/TimestampMillisecondsFormatter.kt
@@ -1,4 +1,4 @@
-package com.gb.stopwatch
+package com.gb.stopwatch.domain
 
 class TimestampMillisecondsFormatter {
