+++ app/src/main/java/com/gb/stopwatch/domain/TimestampProvider.kt
@@ -1,4 +1,4 @@
-package com.gb.stopwatch
+package com.gb.stopwatch.domain
 
 interface TimestampProvider {
     fun getMilliseconds(): Long
