+++ app/src/main/java/com/gb/stopwatch/viewmodel/MainViewModel.kt
@@ -3,7 +3,7 @@ package com.gb.stopwatch.viewmodel
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
-import com.gb.stopwatch.*
+import com.gb.stopwatch.domain.*
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.SupervisorJob
